refactor base_waypoint related logic into WaypointTracker and associated Python module installation, start to use tl_detector to fake traffic light detection.
